Move `PUBLIC_KEY_GUARD` to separate file and add comment
